==> 10-Report-Metrics.ipynb <==
import pandas as pd
import os
import numpy as np

def printStats(data):
    stats = pd.read_pickle(data)
    print(data)
    print(stats)


printStats("results/cnn/GRU-256-Glove-los_3-10-new-cnn-.p")
printStats("results/cnn/GRU-256-Glove-los_7-10-new-cnn-.p")
printStats("results/cnn/GRU-256-Glove-mort_hosp-10-new-cnn-.p")
printStats("results/cnn/GRU-256-Glove-mort_icu-10-new-cnn-.p")

printStats("results/256-GRU-los_3-10-new.p")
printStats("results/256-GRU-los_7-10-new.p")
printStats("results/256-GRU-mort_hosp-10-new.p")
printStats("results/256-GRU-mort_icu-10-new.p")
printStats("results/256-LSTM-los_3-10-new.p")
printStats("results/256-LSTM-los_7-10-new.p")
printStats("results/256-LSTM-mort_hosp-10-new.p")
printStats("results/256-LSTM-mort_icu-10-new.p")


printStats("results/cnn/GRU-256-word2vec-mort_icu-10-new-cnn-.p")
printStats("results/cnn/GRU-256-word2vec-mort_hosp-10-new-cnn-.p")
printStats("results/cnn/GRU-256-word2vec-los_7-10-new-cnn-.p")
printStats("results/cnn/GRU-256-word2vec-los_3-10-new-cnn-.p")
printStats("results/cnn/GRU-256-fasttext-mort_icu-10-new-cnn-.p")
printStats("results/cnn/GRU-256-fasttext-mort_hosp-10-new-cnn-.p")
printStats("results/cnn/GRU-256-fasttext-los_7-10-new-cnn-.p")
printStats("results/cnn/GRU-256-fasttext-los_3-10-new-cnn-.p")
printStats("results/cnn/GRU-256-concat-mort_icu-10-new-cnn-.p")
printStats("results/cnn/GRU-256-concat-mort_hosp-10-new-cnn-.p")
printStats("results/cnn/GRU-256-concat-los_7-10-new-cnn-.p")
printStats("results/cnn/GRU-256-concat-los_3-10-new-cnn-.p")
